requirements.txt
attrs==19.3.0
Automat==0.8.0
blinker==1.4
bluepy==1.3.0
btlewrap==0.0.10
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
miflora==0.7.1
mongoengine==0.23.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
praw==7.5.0
prawcore==2.3.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.1
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==27.4
ufw==0.36
unattended-upgrades==0.1
update-checker==0.18.0
urllib3==1.25.8
wadllib==1.3.3
websocket-client==1.2.3
zipp==1.0.0
zope.interface==4.7.1
